pyproject.toml
[project]
name = "local-bielik"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.10.0",
    "pandas>=2.3.1",
    "protobuf>=6.32.0",
    "pyyaml>=6.0.2",
    "sentence-transformers>=5.1.0",
    "torch>=2.8.0",
    "transformers>=4.55.0",
]
